refactor(runtime): consolidate scheduler constants and hibernation logic
The WASM scheduler (scheduler_wasm.rs) duplicated five scheduling
constants (HEW_MSG_BUDGET, HEW_DEFAULT_REDUCTIONS, HEW_PRIORITY_HIGH,
HEW_PRIORITY_NORMAL, HEW_PRIORITY_LOW) that were already defined as
pub const in actor.rs and imported correctly by the native scheduler.
Remove the local re-definitions and import from crate::actor instead.

Both schedulers also contained identical hibernation-tracking blocks
(~8 lines each) that update idle_count and the hibernating flag after
each activation cycle. Extract this as pub(crate) update_hibernation_state
in actor.rs and replace both inline blocks with a single call site.
The WASM call site casts the local HewActor pointer to crate::actor::HewActor
using the same layout-verified cast already used throughout scheduler_wasm.rs.

No behaviour change: values, branch conditions, and atomic orderings are
unchanged. All 1126 hew-runtime unit tests pass.
